Refactor: Changed weight unit to be a float in code and in db

+++ BaseLibrary/BaseModels/ModelSet.cs
@@ -1,8 +1,8 @@
 namespace ftDB.BaseLibrary.Models
 {
-    public class ModelSet(int weight, int reps, int setNumber)
+    public class ModelSet(double weight, int reps, int setNumber)
     {
-        public int Weight { get; set; } = weight;
+        public double Weight { get; set; } = weight;
         public int Reps { get; set; } = reps;
         public int SetNumber { get; set; } = setNumber;
     }
